fix: remove circular import in peagen
