+++ pkgs/standards/peagen/peagen/core/doe_core.py
@@ -24,7 +24,7 @@ from urllib.parse import urlparse
 from peagen._utils.config_loader import load_peagen_toml
 from peagen.plugins import registry
 from peagen.schemas import DOE_SPEC_V1_1_SCHEMA          # already vendored
-from peagen.cli.commands.validate import _validate           # re-use existing helper
+from peagen._utils._validation import _validate
 
 # ─────────────────────────────── util ──────────────────────────────────────
 _LLM_FALLBACK_KEYS = {
